HBASE-22454 refactor WALSplitter
