+++ hbase-server/src/test/java/org/apache/hadoop/hbase/wal/TestWALMethods.java
@@ -39,7 +39,6 @@ import org.apache.hadoop.hbase.testclassification.RegionServerTests;
 import org.apache.hadoop.hbase.testclassification.SmallTests;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.hadoop.hbase.util.FSUtils;
-import org.apache.hadoop.hbase.wal.WALSplitter.EntryBuffers;
 import org.apache.hadoop.hbase.wal.WALSplitter.PipelineController;
 import org.apache.hadoop.hbase.wal.WALSplitter.RegionEntryBuffer;
 import org.junit.ClassRule;
@@ -89,17 +88,17 @@ public class TestWALMethods {
     Path regiondir = util.getDataTestDir("regiondir");
     fs.delete(regiondir, true);
     fs.mkdirs(regiondir);
-    Path recoverededits = WALSplitter.getRegionDirRecoveredEditsDir(regiondir);
-    String first = WALSplitter.formatRecoveredEditsFileName(-1);
+    Path recoverededits = WALSplitUtil.getRegionDirRecoveredEditsDir(regiondir);
+    String first = WALSplitUtil.formatRecoveredEditsFileName(-1);
     createFile(fs, recoverededits, first);
-    createFile(fs, recoverededits, WALSplitter.formatRecoveredEditsFileName(0));
-    createFile(fs, recoverededits, WALSplitter.formatRecoveredEditsFileName(1));
-    createFile(fs, recoverededits, WALSplitter
+    createFile(fs, recoverededits, WALSplitUtil.formatRecoveredEditsFileName(0));
+    createFile(fs, recoverededits, WALSplitUtil.formatRecoveredEditsFileName(1));
+    createFile(fs, recoverededits, WALSplitUtil
         .formatRecoveredEditsFileName(11));
-    createFile(fs, recoverededits, WALSplitter.formatRecoveredEditsFileName(2));
-    createFile(fs, recoverededits, WALSplitter
+    createFile(fs, recoverededits, WALSplitUtil.formatRecoveredEditsFileName(2));
+    createFile(fs, recoverededits, WALSplitUtil
         .formatRecoveredEditsFileName(50));
-    String last = WALSplitter.formatRecoveredEditsFileName(Long.MAX_VALUE);
+    String last = WALSplitUtil.formatRecoveredEditsFileName(Long.MAX_VALUE);
     createFile(fs, recoverededits, last);
     createFile(fs, recoverededits,
       Long.toString(Long.MAX_VALUE) + "." + System.currentTimeMillis());
@@ -108,21 +107,21 @@ public class TestWALMethods {
     FSUtils.setRootDir(walConf, regiondir);
     (new WALFactory(walConf, "dummyLogName")).getWAL(null);
 
-    NavigableSet<Path> files = WALSplitter.getSplitEditFilesSorted(fs, regiondir);
+    NavigableSet<Path> files = WALSplitUtil.getSplitEditFilesSorted(fs, regiondir);
     assertEquals(7, files.size());
     assertEquals(files.pollFirst().getName(), first);
     assertEquals(files.pollLast().getName(), last);
     assertEquals(files.pollFirst().getName(),
-      WALSplitter
+        WALSplitUtil
         .formatRecoveredEditsFileName(0));
     assertEquals(files.pollFirst().getName(),
-      WALSplitter
+        WALSplitUtil
         .formatRecoveredEditsFileName(1));
     assertEquals(files.pollFirst().getName(),
-      WALSplitter
+        WALSplitUtil
         .formatRecoveredEditsFileName(2));
     assertEquals(files.pollFirst().getName(),
-      WALSplitter
+        WALSplitUtil
         .formatRecoveredEditsFileName(11));
   }
